add button on MainActivity and make ImportMenuActivity and OwnerActivity and Split folder
