+++ .idea/misc.xml
@@ -1,5 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="..\:/Users/msi/Desktop/Projects/Euphony_Listen_My_Order/listen-my-order/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/Users/msi/Desktop/Projects/Euphony_Listen_My_Order/listen-my-order/app/src/main/res/layout/activity_owner.xml" value="0.1" />
+      </map>
+    </option>
+  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
