+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.listen_my_order">
+
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
 
     <application
@@ -11,7 +12,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Listenmyorder">
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.ImportMenuActivity"
+            android:exported="true" />
+        <activity
+            android:name=".activities.OwnerActivity"
+            android:exported="true" />
+        <activity
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
