+++ app/src/main/java/com/example/listen_my_order/activities/MainActivity.java
@@ -1,25 +1,31 @@
-package com.example.listen_my_order;
+package com.example.listen_my_order.activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
 
 import android.Manifest;
+import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import com.example.listen_my_order.R;
 
 import euphony.lib.receiver.EuRxManager;
-import euphony.lib.transmitter.EuPlayer;
 
 public class MainActivity extends AppCompatActivity {
+    private Button importMenuButton;
+    private Button ownerButton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        EuRxManager rxManager = new EuRxManager();
-        rxManager.listen();
+        importMenuButton = findViewById(R.id.import_menu_button);
+        ownerButton = findViewById(R.id.owner_button);
 
         int permission = ContextCompat.checkSelfPermission(MainActivity.this, Manifest.permission.RECORD_AUDIO);
 
@@ -29,6 +35,22 @@ public class MainActivity extends AppCompatActivity {
                         Manifest.permission.RECORD_AUDIO}, 1000);
             }
         }
+
+        importMenuButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                startActivity(new Intent(MainActivity.this, ImportMenuActivity.class));
+            }
+        });
+
+        ownerButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                startActivity(new Intent(MainActivity.this, OwnerActivity.class));
+            }
+        });
+
+
     }
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
