+++ app/src/main/java/com/example/listen_my_order/activities/OwnerActivity.java
@@ -1,0 +1,16 @@
+package com.example.listen_my_order.activities;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.listen_my_order.R;
+
+public class OwnerActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_owner);
+    }
+}
